clipproperties: Allow locking the aspect ratio
Currently, the clip properties panel supports changing the size of clips
using height and width spin buttons.

In GIMP there is a function to lock this ratio while resizing, but in
Pitivi users had to calculate it themselves.

We added a button similar to the one in GIMP to the clip properties
panel.

The svg files were pulled directly from Gimp's files.
https://gitlab.gnome.org/GNOME/gimp/-/blob/master/icons/Symbolic/scalable/gimp-vchain-broken-symbolic.svg
https://gitlab.gnome.org/GNOME/gimp/-/blob/master/icons/Symbolic/scalable/gimp-vchain-symbolic.svg

Fixes #2446