projectsettings: Remove the Audio widgets
The most important for a (video) project is the frame rate. The audio
rate is something which is there and can be super easily changed. That's
why we don't bother with it in the project settings.

We can afford this because TL/DR the audio settings are initialized when
importing the first (audio) clip.

The audio settings do matter when rendering, and the user can change
them there. In this case the project will remain with the audio settings
used in the Render dialog.

Fixes #2559