timeline: Fix selection post grouping/ungrouping
Previously, if you selected an a/v clip and ungroup it, then reselect
the two resulting clips and re-group them, the resulting clip would
appear with the video selected and audio deselected.

Previously, if you ungrouped a group/clip, the selection would reset.

Now the selection persists grouping and ungrouping.