clipproperties: Add a compositing expander
Pitivi allows users to manually keyframe the opacity of a clip to create
fade-in and fade-out transitions.

However, this is relatively complex for basic use cases, and can be
tedious to do on a large scale.

This commit adds a compositing expander to the clip properties panel, in
which the user can easily set fade-in and fade-out durations for a clip.

Fixes #1472