ruler: Fix ticks disappearing
We used to break the drawing of ticks when we encounter ticks which have
space between them less than MIN_TICK_SPACING_PIXELS. But now we are
iterating from most frequent to rare ticks. So the 'break' is now changed
to 'continue' so that it won't completely break the drawing loop.

Fixes #2622